Use a new rewiter, instead of adding inheritdoc to many test outputs
