+++ src/PropertyChanged.SourceGenerator.UnitTests/NullableTypeTests.DoesNotGenerateNullableEventIfInFileNullableContext_SomeViewModel.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class SomeViewModel : global::System.ComponentModel.INotifyPropertyChanged
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangedEventHandler PropertyChanged;
     #nullable enable annotations
     public string Foo { get; set; }
