+++ src/PropertyChanged.SourceGenerator.UnitTests/NullableTypeTests.GeneratesNullableEventIfInCompilationNullableContext_SomeViewModel.verified.cs
@@ -1,7 +1,6 @@
 ﻿#nullable enable
 partial class SomeViewModel : global::System.ComponentModel.INotifyPropertyChanged
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangedEventHandler? PropertyChanged;
     public string Foo { get; set; }
     protected virtual void OnPropertyChanged(global::System.ComponentModel.PropertyChangedEventArgs eventArgs)
