+++ src/PropertyChanged.SourceGenerator.UnitTests/OnAnyPropertyChangedTests.DoesNotGenerateForOverriddenMethod_Base.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class Base : global::System.ComponentModel.INotifyPropertyChanged
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangedEventHandler PropertyChanged;
     public string Foo { get; set; }
     protected virtual void OnPropertyChanged(global::System.ComponentModel.PropertyChangedEventArgs eventArgs)
