+++ src/PropertyChanged.SourceGenerator.UnitTests/OnAnyPropertyChangedTests.cs
@@ -15,7 +15,7 @@ public class OnAnyPropertyChangedTests : TestsBase
 {
     private static readonly CSharpSyntaxVisitor<SyntaxNode?>[] rewriters = new CSharpSyntaxVisitor<SyntaxNode?>[]
     {
-        RemovePropertiesRewriter.Instance, RemoveInpcMembersRewriter.CommentsOnly
+        RemovePropertiesRewriter.Instance, RemoveDocumentationRewriter.Instance,
     };
 
     [Test]
