+++ src/PropertyChanged.SourceGenerator.UnitTests/OnAnyPropertyChangingTests.DoesNotGenerateForOverriddenMethod_Base.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class Base
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangingEventHandler PropertyChanging;
     public string Foo { get; set; }
     protected virtual void OnPropertyChanging(global::System.ComponentModel.PropertyChangingEventArgs eventArgs)
