+++ src/PropertyChanged.SourceGenerator.UnitTests/OnAnyPropertyChangingTests.GeneratesParameterlessWhenAlsoGeneratingRaisePropertyChanging_SomeViewModel.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class SomeViewModel
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangingEventHandler PropertyChanging;
     public string Foo { get; set; }
     protected virtual void OnPropertyChanging(global::System.ComponentModel.PropertyChangingEventArgs eventArgs)
