+++ src/PropertyChanged.SourceGenerator.UnitTests/RaisePropertyChangedCallTests.DoesNotGenerateInpcInterfaceIfAlreadySpecified_SomeViewModel.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class SomeViewModel
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangedEventHandler PropertyChanged;
     public string Foo { get; set; }
     protected virtual void OnPropertyChanged(global::System.ComponentModel.PropertyChangedEventArgs eventArgs)
