+++ src/PropertyChanged.SourceGenerator.UnitTests/RaisePropertyChangedDefinitionTests.HandlesSealedClass_SomeViewModel.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class SomeViewModel : global::System.ComponentModel.INotifyPropertyChanged
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangedEventHandler PropertyChanged;
     public string Foo { get; set; }
     private void OnPropertyChanged(global::System.ComponentModel.PropertyChangedEventArgs eventArgs)
