+++ src/PropertyChanged.SourceGenerator.UnitTests/RaisePropertyChangingCallTests.TakesNameFromPropertyChangedIfPossible_SomeViewModel.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class SomeViewModel
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangingEventHandler PropertyChanging;
     public int Foo { get; set; }
     protected virtual void RaisePropertyChanging(global::System.ComponentModel.PropertyChangingEventArgs eventArgs)
