+++ src/PropertyChanged.SourceGenerator.UnitTests/RaisePropertyChangingCallTests.cs
@@ -15,7 +15,7 @@ public class RaisePropertyChangingCallTests : TestsBase
 {
     private static readonly CSharpSyntaxVisitor<SyntaxNode?>[] rewriters = new CSharpSyntaxVisitor<SyntaxNode?>[]
     {
-        RemoveInpcMembersRewriter.Changed, RemovePropertiesRewriter.Instance
+        RemoveInpcMembersRewriter.Changed, RemovePropertiesRewriter.Instance, RemoveDocumentationRewriter.Instance,
     };
 
     [Test]
