+++ src/PropertyChanged.SourceGenerator.UnitTests/RaisePropertyChangingDefinitionTests.DefinesVirtual_Derived.verified.cs
@@ -1,6 +1,5 @@
 ﻿partial class Derived
 {
-    /// <inheritdoc />
     public event global::System.ComponentModel.PropertyChangingEventHandler PropertyChanging;
     public string Bar { get; set; }
     protected virtual void OnPropertyChanging(global::System.ComponentModel.PropertyChangingEventArgs eventArgs)
